% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactPotentialProfiling_AACPMatrix.R
\name{ContactPotentialProfiling_AACPMatrix}
\alias{ContactPotentialProfiling_AACPMatrix}
\alias{CPP_AACPMatrix}
\title{Amino acid pairwise contact potential (AACP) matrix.}
\usage{
CPP_AACPMatrix(sourceFile = system.file("AACPMatrix.csv", package =
  "Repitope"))
}
\arguments{
\item{sourceFile}{A source file. By default, the AACP dataset downloaded from the official AAIndex database will be used. Alternatively, users can provide modified AACP scales.}
}
\description{
\code{CPP_AACPMatrix} generates a formatted AACP datatable.
}
