% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_EpitopeDataset.R
\name{DataPreparation_EpitopeDataset}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Import}
\alias{compressedToLongFormat}
\alias{compressedToDummyDF}
\title{Compile epitope datasets retrieved from IEDB and other sources.}
\usage{
Epitope_Import(IEDBAssayFileName = NULL, OtherFileNames = NULL,
  peptideLengthSet = 8:11)

compressedToLongFormat(df, compressedColumnName)

compressedToDummyDF(df, compressedColumnName)
}
\arguments{
\item{IEDBAssayFileName}{A set of T cell assay results obtained from IEDB.}

\item{OtherFileNames}{Other epitope source files. Must contain a "Dataset" column indicating the data source.}

\item{peptideLengthSet}{A set of peptide lengths to be incorporated.}

\item{df}{A dataframe/datatable of epitope dataset.}

\item{compressedColumnName}{A string indicating the names of the compressed column to be converted into a long format.}
}
\description{
\code{Epitope_Import} imports IEDB and other epitope files. Contradicting annotations on immunogenicity are internally resolved.\cr
\code{compressedToLongFormat} converts a dataframe with a compresed column in which items are concatenated with "|" into a long-format one.\cr
\code{compressedToDummyDF} generates dummy variable columns from a compresed column in which items are concatenated with "|".\cr
}
