% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_NetMHC.R
\name{DataPreparation_NetMHC}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_PeptideExport}
\alias{NetMHC_Script}
\alias{NetMHC_Import}
\title{Utility functions for NetMHC.}
\usage{
NetMHC_PeptideExport(peptideSet, outDir = "./NetMHC/",
  maxPeptides = NULL)

NetMHC_Script(outDir = "./NetMHC/")

NetMHC_Import(netMHCOutputFileNames, MHCType = c("MHCI", "MHCII"))
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{outDir}{A path to the output directory in which peptides will be exported in FASTA format.}

\item{maxPeptides}{An integer indicating the maximum number of peptides in each FASTA files. Set \code{NULL} to disable chunking.}

\item{netMHCOutputFileNames}{A set of tab-delimited NetMHC result files.}

\item{MHCType}{A character string. "MHCI" or "MHCII".}
}
\description{
To use the scripts generated by \code{NetMHC_Script}, NetMHC4.0 and NetMHCIIpan3.2 should be installed locally.
}
