% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_TCR_FastQDump.R
\name{DataPreparation_TCR_FastQDump}
\alias{DataPreparation_TCR_FastQDump}
\alias{FastQDump_Script}
\title{Downloading fastQ files from NCBI Sequence Read Archive (SRA).}
\usage{
FastQDump_Script(accessionSet, outDir = "C:/SRA/output",
  FastQDumpPath = "C:\\\\SRA\\\\sratoolkit.2.9.2-win64\\\\bin\\\\fastq-dump.exe")
}
\arguments{
\item{accessionSet}{A set of SRA Run Accession numbers.}

\item{outDir}{A directory for outputs.}

\item{FastQDumpPath}{A full path to fastq-dump.exe.}
}
\description{
\code{FastQDump_Script} generates fustq-dump script.
}
