% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_TCRPeptidePairDataset.R
\docType{data}
\name{Dataset_TCRPeptidePairDataset}
\alias{Dataset_TCRPeptidePairDataset}
\alias{TCRPeptidePairDataset}
\title{A compiled dataset of human TCR-peptide pairs.}
\format{data.table}
\usage{
TCRPeptidePairDataset
}
\description{
\code{TCRPeptidePairDataset} is a set of epitopes and cognate TCR sequences, with affinity and structural annotations
}
\keyword{datasets}
