% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_ExtractFeatures.R
\name{Features_ExtractFeatures}
\alias{Features_ExtractFeatures}
\alias{Features_ExtractParameters}
\title{Extract features from a list of feature dataframes.}
\usage{
Features_ExtractFeatures(preprocessedDFList, criteria = "intersect")

Features_ExtractParameters(preprocessedDFList, criteria = "intersect")
}
\arguments{
\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess}, \code{Features_CorFilter}, \code{Features_FeatureSelect}, \code{Features_Importance}, or \code{Features_Importance_Reduce}.}

\item{criteria}{The criteria of feature extraction. Can either be "intersect" or "union".}
}
\description{
\code{Features_ExtractFeatures} obtains a set of features from a list of preprocessed/feature-selected dataframes. It also generates a venn diagram.\cr
\code{Features_ExtractParameters} obtains a set of parameters.\cr
}
