% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_IO.R
\name{Features_IO}
\alias{Features_IO}
\alias{readFeatureDFList}
\alias{saveFeatureDFList}
\title{Importing and exporting utilities for feature dataframes.}
\usage{
readFeatureDFList(featureDFFileNames)

saveFeatureDFList(featureDFList, fileNameHeader)
}
\arguments{
\item{featureDFFileNames}{A character vector of binary file names with either ".fst" or ".rds" extensions.}

\item{featureDFList}{A named list of feature dataframes.}

\item{fileNameHeader}{A file name header.}
}
\description{
Importing and exporting utilities for feature dataframes.
}
