% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunogenicity.R
\name{Immunogenicity}
\alias{Immunogenicity}
\alias{Immunogenicity_TrainModels}
\alias{Immunogenicity_SummarizeInternalScores}
\alias{Immunogenicity_Score}
\alias{Immunogenicity_Score_Cluster}
\alias{Immunogenicity_Predict}
\title{Immunogenicity score.}
\usage{
Immunogenicity_TrainModels(featureDF, metadataDF, featureSet = "all",
  seedSet = 1:5, coreN = parallel::detectCores(logical = F))

Immunogenicity_SummarizeInternalScores(trainModelResults)

Immunogenicity_Score(featureDF, metadataDF, featureSet = "all",
  seedSet = 1:5, coreN = parallel::detectCores(logical = F))

Immunogenicity_Score_Cluster(featureDF, metadataDF, clusterDF,
  featureSet = "all", seedSet = 1:5,
  coreN = parallel::detectCores(logical = F))

Immunogenicity_Predict(externalFeatureDFList, trainModelResults)
}
\arguments{
\item{featureDF}{A dataframe of features generated by \code{Features}.}

\item{metadataDF}{A dataframe containing metadata, consisting of "Peptide" and "Immunogenicity" columns. Other columns if provided would be used as features.}

\item{featureSet}{A set of features used for model training. Set "all" to shortcut.}

\item{seedSet}{A set of random seeds.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to disable.}

\item{trainModelResults}{The model training result returned by \code{Immunogenicity_TrainModels}.}

\item{clusterDF}{A dataframe containing cluster metadata, consisting of "Peptide" and "Cluster" columns.}

\item{externalFeatureDFList}{The feature dataframes for which immunogenicity is to be predicted.}
}
\description{
\code{Immunogenicity_TrainModels} performes preprocessing, trains extremely randomized tree models, and internally predicts immunogenicity scores on the entire dataset provided.\cr
\code{Immunogenicity_SummarizeInternalScores} summarizes internally calculated immunogenicity scores.\cr
\code{Immunogenicity_Score} is a wrapper function of \code{Immunogenicity_TrainModels} and \code{Immunogenicity_SummarizeInternalScores}.\cr
\code{Immunogenicity_Score_Cluster} is similar to \code{Immunogenicity_Score}, but takes \code{clusterDF} to filter peptides based on their similarities.\cr
\code{Immunogenicity_Predict} predicts immunogenicity scores on external datasets. Models trained by \code{Immunogenicity_TrainModels} are necessary. Note that, due to Java constraints, models have to be constructed in each new R session.
}
