% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborNetwork_Clustering.R
\name{NeighborNetwork_Clustering}
\alias{NeighborNetwork_Clustering}
\alias{neighborNetwork_ConnectedSubGraphDF}
\alias{neighborNetwork_ConnectedSubGraph}
\alias{neighborNetwork_Cluster}
\alias{neighborNetwork_Cluster_Batch}
\alias{neighborNetwork_Cluster_FeatureDF}
\title{Neighbor network clustering analysis.}
\usage{
neighborNetwork_ConnectedSubGraphDF(neighborNetResult, peptide)

neighborNetwork_ConnectedSubGraph(neighborNetResult, peptide)

neighborNetwork_Cluster(peptide, graph, metadataDF, seed = 12345,
  plot = T)

neighborNetwork_Cluster_Batch(neighborNetResult, metadataDF,
  seed = 12345, coreN = parallel::detectCores(logical = F))

neighborNetwork_Cluster_FeatureDF(neighborNetClusterResult,
  coreN = parallel::detectCores(logical = F))
}
\arguments{
\item{neighborNetResult}{The result returned from \code{neighborNetwork}.}

\item{peptide}{A target peptide sequence.}

\item{graph}{A directed and weighted neighbor network of the target peptide.}

\item{metadataDF}{A dataframe which has "Peptide" and "ImmunogenicityScore" columns. Optionally, an "Immunogenicity" column would be integrated if present.}

\item{seed}{A random seed.}

\item{plot}{Logical. Whether the network cluster plot shuould be generated.}

\item{coreN}{The number of threads for parallelization.}

\item{neighborNetClusterResult}{The result returned from \code{neighborNetwork_Cluster_Batch}.}
}
\description{
\code{neighborNetwork_ConnectedSubGraph} and \code{neighborNetwork_ConnectedSubGraphDF} extract the minimum connected subgraph focusing on a target peptide.\cr
\code{neighborNetwork_Cluster} and \code{neighborNetwork_Cluster_Batch} conduct a network clustering analysis using a walktrap algorithm.\cr
\code{neighborNetwork_Cluster_FeatureDF} computes per-cluster and per-peptide features, including "escape potential".
}
