% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTCRAnalysis.R
\name{contactFootprintDensityPlot}
\alias{contactFootprintDensityPlot}
\alias{Features_sCPP}
\alias{univariateCorrelationAnalysis}
\alias{univariateCorrelationPlot}
\alias{univariateFeatureSelect}
\alias{multivariateRegressionPlot}
\alias{multivariateRegressionPrediction}
\title{Single TCR-peptide contact potential profiling.}
\usage{
contactFootprintDensityPlot(weightStringSet)

Features_sCPP(peptideSet, tcrSet, aaIndexIDSet = c("BETM990101inv",
  "KESO980102inv", "KOLA930101", "MICC010101", "SIMK990104", "VENM980101",
  "VENM980101inv"), fragLenSet = 3:8, seed = 12345,
  coreN = parallel::detectCores(logical = F))

univariateCorrelationAnalysis(dt_deltaG_sCPP,
  coreN = parallel::detectCores(logical = F))

univariateCorrelationPlot(dt_deltaG_sCPP, dt_deltaG_sCPP_outlier = NULL,
  corVarName = "sCPP_BETM990101inv_Skew_4")

univariateFeatureSelect(dt_univ, aaIndexIDSet, sig = 0.05,
  bestOnly = T)

multivariateRegressionPlot(dt_deltaG_sCPP, dt_deltaG_sCPP_outlier = NULL,
  corVarNames)

multivariateRegressionPrediction(dt_deltaG_sCPP,
  dt_deltaG_sCPP_new = NULL, corVarNames)
}
\arguments{
\item{weightStringSet}{A set of contact weights. E.g. "1|2|3|2|1"}

\item{peptideSet}{A set of peptide sequences.}

\item{tcrSet}{A set of corresponding TCR sequences.}

\item{aaIndexIDSet}{A set of AAIndex IDs indicating the AACP scales to be used. Set "all" to shortcut the selection of all available AACP scales.}

\item{fragLenSet}{A set of sliding window sizes. Must be between 3 and 8.}

\item{seed}{A random seed.}

\item{coreN}{The number of cores to be used for parallelization.}

\item{dt_deltaG_sCPP}{A data.table containing the "DeltaG" and sCPP feature columns.}

\item{dt_deltaG_sCPP_outlier}{Optional. If provided, it would be used as outliers.}

\item{corVarName}{A variable name used for univariate analysis.}

\item{dt_univ}{An output data.table of univariate analysis.}

\item{sig}{A siginificance cutoff for univariate feature selection.}

\item{bestOnly}{Logical. If True, only the best sCPP feature per AAIndex will be returned. If False, all sCPP features derived from selected AAIndices will be returned.}

\item{corVarNames}{A set of variable names used for multivariate regression.}

\item{dt_deltaG_sCPP_new}{Optional. If provided, TCR affnities would be predicted for this dataset. Otherwise, predicted TCR affinities for the original dataset would be returned.}
}
\description{
Single TCR-peptide contact potential profiling.
}
