% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_DF.R
\name{Utility_DF}
\alias{Utility_DF}
\alias{sortByMedian}
\alias{mutate_if_byrow}
\title{Utility functions for data frames.}
\usage{
sortByMedian(df, groupColumnName = "FeatureID",
  valueColumnName = "Importance", descending = F)

mutate_if_byrow(df, .predictate, ...)
}
\arguments{
\item{df}{Any dataframe.}

\item{groupColumnName}{The column name for the grouping variables.}

\item{valueColumnName}{The column name for values.}

\item{descending}{Logical. Whether the dataframe should be reordered in a descending order.}

\item{.predictate}{A predictate function to be applied to the rows or a logical vector. Note: columns to be mutated have to be created in advance.}

\item{...}{Other parameters passed to \code{dplyr::mutate}.}
}
\description{
Utility functions for data frames.
}
